11.10.2024: Arreglo rutas, método que impide duplicados de clientes, método que consulta disponibilidad de una vivienda

+++ src/api/controllers/Feature.cjs
@@ -3,6 +3,8 @@ const { returnMessage } = require("../../utils/returnMessage.cjs");
 // Importar el modelo
 const Feature = require("../models/Feature.cjs");
 
+const { deleteImageFromCloudinary } = require("../../utils/deleteImageFromCloudinary.cjs");
+
 // Función que lista los registros de la colección
 const getFeatures = async (req, res, next) => {
     try {
@@ -38,6 +40,12 @@ const postFeature = async (req, res, next) => {
         // Crear la variable que contendrá los datos
         const newFeature = new Feature(req.body);
 
+        // Comprobar si hay archivo req.file
+        if(req.file) {
+            // Asignar el nombre del archivo al campo image de la colección
+            newFeature.icon = req.file.path;
+        }
+
         // Guardar el nuevo registro en la base de datos
         const feature = await newFeature.save();
 
@@ -49,42 +57,33 @@ const postFeature = async (req, res, next) => {
     }
 }
 
-// Función que actualiza un registro en la colección
-const putFeature = async (req, res, next) => {
-    try {
-        // Recoger el id del registro a actualizar
-        const { id } = req.params;
-
-        // Recoger los datos del registros a actualizar
-        const newFeature = new Feature(req.body);
-        
-        // Poner mismo id al registro
-        newFeature._id = id;
-
-        // Actualizar registro en la base de datos
-        const featureUpdated = await Feature.findByIdAndUpdate(id, newFeature, { new: true });
-
-        // Devolver resultado OK y el registro actualizado
-        returnMessage(res, 200, "Registro actualizado con éxito", featureUpdated);
-    } catch (error) {
-        returnMessage(res, 400, "Error al actualizar el registro");
-    }
-}
 // Función que elimina un registro de la colección
 const deleteFeature = async (req, res, next) => {
     try {
-        // Recoger el id del registro a eliminar
         const { id } = req.params;
+        const featureToDelete = await Feature.findById(id);
+    
+        if (!featureToDelete) {
+          return returnMessage(res, 404, "Registro no encontrado");
+        }
+    
+        // Verificar si la imagen no es la default antes de eliminarla de Cloudinary
+        if (featureToDelete.icon !== "default-icon.jpg") {
 
-        // Eliminar el registro de la BBDD
+            // Obtener el publicId de la imagen para poder eliminarla
+            // El publicId es el nombre de la imagen sin su extensión
+            const publicId = `hosty_house/icons/${featureToDelete.icon.split('/').pop().split('.')[0]}`;        
+
+            // Intentar eliminar la imagen de Cloudinary
+            deleteImageFromCloudinary(res, publicId);
+        }
+    
         const featureDeleted = await Feature.findByIdAndDelete(id);
-
-        // Devolver resultado OK y el registro eliminado
         returnMessage(res, 200, "Registro eliminado con éxito", featureDeleted);
-    } catch (error) {
+      } catch (error) {
         returnMessage(res, 400, "Error al eliminar el registro");
-    }
+      }
 }
 
 // Exportar funciones del controlador
-module.exports = { getFeatures, getFeature, postFeature, putFeature, deleteFeature }
+module.exports = { getFeatures, getFeature, postFeature, deleteFeature }
